Abstract
In Major League Baseball (MLB), injuries to pitchers are regular occurrences. Occasionally, pitchers returning to play after a stint on the injured list (IL) may not be fully recovered and potentially require a second stint to further rehabilitate the original injury. While some research has been done in monitoring pitcher performance or throwing mechanics as an indicator of possible re-injury, the retrospective nature of these methods may not necessarily promote expediency in re-injury identification. Thus, the purpose of the present study is to propose and evaluate a control charting technique, which are traditionally used in quality control manufacturing applications, for monitoring pitcher performance in real time. If a pitcher’s performance is deemed “out-of-control,” then it is plausible that such a performance may be indicative of lingering injury. The development of the method is reviewed, and its efficacy is evaluated. Practical implications for application and limitations of the method are also discussed.